+++ Disciple/src/utopia/disciple/model/request/FileBody.scala
@@ -1,7 +1,10 @@
 package utopia.disciple.model.request
 
 import utopia.access.model.ContentType
-import java.io.{File, FileInputStream}
+import utopia.flow.parse.AutoClose._
+import utopia.flow.parse.StreamExtensions._
+
+import java.io.{File, FileInputStream, OutputStream}
 import scala.util.Try
 
 /**
@@ -9,13 +12,26 @@ import scala.util.Try
 * @author Mikko Hilpinen
 * @since 1.5.2018
 **/
-class FileBody(val file: File, val contentType: ContentType, val contentEncoding: Option[String] = None) extends Body
+class FileBody(file: File, override val contentType: ContentType, override val contentEncoding: Option[String] = None)
+	extends RequestBody
 {
-	override def repeatable = true
-	override def chunked = false
+	// ATTRIBUTES   ---------------------
 	
-    override def contentLength = Some(file.length())
-    override def stream = Try(new FileInputStream(file))
+	private val len = file.length()
+	
+	override val repeatable = true
+	override val streaming: Boolean = true
+	
+	
+	// IMPLEMENTED  --------------------
+	
+	override def contentLength = Some(len)
 	
 	override def toString = s"${file.getName} ($contentType)"
+	
+	override def writeTo(output: OutputStream): Try[Unit] = Try {
+		new FileInputStream(file).consume { _.writeTo(output) }
+	}
+	
+	override def close(): Unit = ()
 }